requirements.txt
fastapi[standard]~=0.123.4
asyncpg~=0.31.0
sqlalchemy[asyncio]~=2.0.44
python-dotenv~=1.2.1
celery[redis]~=5.6.0
redis~=7.1.0